requirements.txt
flask
Werkzeug
Pillow
pytesseract
opencv-python
numpy
scikit-learn